refactor: 폴더구조 변경, type 명시, coupon 분리
