+++ src/basic/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useDebounce';
 export * from './useNotification';
 export * from './useProduct';
 export * from './useCart';
+export * from './useCoupon';
